[25] Retrieve lost frontend files from older branches due to merge & minor clean up of files and directories
