+++ .FrontEnd_backup/views/pages/demo.ejs
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html lang="en-us">
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+    <title>Unity WebGL Player | Game</title>
+    <link rel="shortcut icon" href="TemplateData/favicon.ico">
+    <link rel="stylesheet" href="TemplateData/style.css">
+    <script src="TemplateData/UnityProgress.js"></script>
+    <script src="Build/UnityLoader.js"></script>
+    <script>
+        var gameInstance = UnityLoader.instantiate("gameContainer", "Build/WebGL_demoGame.json", {
+            onProgress: UnityProgress
+        });
+    </script>
+</head>
+
+<div id="logout">
+    <form action='/logout' onsubmit="return confirm('Are you sure you want to log out?')">
+        <input type='submit' value='Logout' , id="logoutBtn" style="width:10%;padding:10px 20px;margin: 2px 1px;font-size:15px;cursor:pointer;" />
+    </form>
+</div>
+
+<iframe src='chatbox' width="300" height="700"></iframe>
+
+<body>
+    <div class="webgl-content">
+        <div id="gameContainer" style="width: 960px; height: 600px"></div>
+        <div class="footer">
+            <div class="webgl-logo"></div>
+            <div class="fullscreen" onclick="gameInstance.SetFullscreen(1)"></div>
+            <div class="title">Game</div>
+        </div>
+    </div>
+</body>
+
+</html>
