+++ .gitignore
@@ -1,2 +1,13 @@
-node_modules/
-coverage/
+# Node build artifacts
+node_modules
+coverage
+npm-debug.log
+
+# Local development
+*.env
+*.dev
+.DS_Store
+
+# Docker
+Dockerfile
+docker-compose.yml
