--- FrontEnd/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "node-js-getting-started",
-  "version": "0.3.0",
-  "description": "A sample Node.js app using Express 4",
-  "engines": {
-    "node": "12.x"
-  },
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "test": "node test.js"
-  },
-  "dependencies": {
-    "alert-node": "^2.0.1",
-    "axios": "^0.19.0",
-    "ejs": "^2.5.6",
-    "express": "^4.17.1",
-    "express-session": "^1.17.0",
-    "mongodb": "^3.3.3",
-    "nodemon": "^1.19.4",
-    "socket.io": "^2.3.0"
-  },
-  "devDependencies": {
-    "request": "^2.81.0",
-    "tape": "^4.7.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/heroku/node-js-getting-started"
-  },
-  "keywords": [
-    "node",
-    "heroku",
-    "express"
-  ],
-  "license": "MIT"
-}
