+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "game",
   "version": "1.0.0",
-  "description": "",
+  "description": "Dungeon Crawler Unity game with Node.js and Express.js",
   "jest": {
     "testURL": "http://localhost/"
   },
@@ -28,7 +28,8 @@
     "mongoose": "^5.7.12",
     "nodemon": "^1.19.4",
     "passport": "^0.4.0",
-    "passport-google-oauth20": "^2.0.0"
+    "passport-google-oauth20": "^2.0.0",
+    "socket.io": "^2.3.0"
   },
   "devDependencies": {
     "jest": "^24.9.0",
@@ -38,7 +39,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/heroku/node-js-getting-started"
+    "url": "https://github.com/web-game-dev/OnlineGame.git"
   },
   "keywords": [
     "node",
