+++ public/css/chatbox.css
@@ -1,0 +1,50 @@
+chatbox {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    font: 13px Helvetica, Arial;
+}
+
+form {
+    background: #fff;
+    padding: 3px;
+    position: fixed;
+    bottom: 0;
+    width: 15%;
+    border-color: #000;
+    border-top-style: solid;
+    border-top-width: 1px;
+}
+
+form input {
+    border-style: solid;
+    border-width: 1px;
+    padding: 10px;
+    width: 100%;
+    margin-right: .5%;
+}
+
+form button {
+    width: 100%;
+    background: rgb(130, 224, 255);
+    border: none;
+    padding: 10px;
+    margin-left: 2%;
+}
+
+#messages {
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+}
+
+#messages li {
+    padding: 5px 10px;
+}
+
+#messages li:nth-child(odd) {
+    background: #eee;
+}
