+++ views/pages/demo.ejs
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html lang="en-us">
-  <head>
+
+<head>
     <meta charset="utf-8">
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
     <title>Unity WebGL Player | Game</title>
@@ -9,26 +10,33 @@
     <script src="TemplateData/UnityProgress.js"></script>
     <script src="Build/UnityLoader.js"></script>
     <script>
-      var gameInstance = UnityLoader.instantiate("gameContainer", "Build/TestWeb.json", {onProgress: UnityProgress});
+        var gameInstance = UnityLoader.instantiate("gameContainer", "Build/WebGL_demoGame.json", {
+            onProgress: UnityProgress
+        });
     </script>
-  </head>
-  
-  <!--Sample on how to display data from the back-end  -->
-  <!-- <h1><%= name%></h1> -->
-  
-  <form action='/logout' onsubmit="return confirm('Are you sure you want to log out?')">
-    <input type='submit' value='Logout' id='logoutBtn'
-    style="width:10%;padding:10px 20px;margin: 2px 1px;font-size:15px;cursor:pointer;"/>
-  </form>
+</head>
 
-  <body>
+<div id="logout">
+    <form action='/logout' onsubmit="return confirm('Are you sure you want to log out?')">
+        <input type='submit' value='Logout' , id="logoutBtn"
+        style="width:10%;padding:10px 20px;margin: 2px 1px;font-size:15px;cursor:pointer;"/>
+    </form>
+</div>
+
+<!--Sample on how to display data from the back-end  -->
+<!-- <h1><%= name%></h1> -->
+
+<iframe src='chatbox' width="300" height="700"></iframe>
+
+<body>
     <div class="webgl-content">
-      <div id="gameContainer" style="width: 960px; height: 600px"></div>
-      <div class="footer">
-        <div class="webgl-logo"></div>
-        <div class="fullscreen" onclick="gameInstance.SetFullscreen(1)"></div>
-        <div class="title">Game</div>
-      </div>
+        <div id="gameContainer" style="width: 960px; height: 600px"></div>
+        <div class="footer">
+            <div class="webgl-logo"></div>
+            <div class="fullscreen" onclick="gameInstance.SetFullscreen(1)"></div>
+            <div class="title">Game</div>
+        </div>
     </div>
-  </body>
+</body>
+
 </html>
